BookMe.Application/Service/Commands/DeleteService/DeleteServiceCommandHandler.cs
using BookMe.Domain.Interfaces;
using MediatR;
using System.Threading;
using System.Threading.Tasks;

namespace BookMe.Application.Service.Commands.DeleteService
{
    public class DeleteServiceCommandHandler : IRequestHandler<DeleteServiceCommand>
    {
        private readonly IServiceRepository _serviceRepository;

        public DeleteServiceCommandHandler(IServiceRepository serviceRepository)
        {
            _serviceRepository = serviceRepository;
        }

        public async Task<Unit> Handle(DeleteServiceCommand request, CancellationToken cancellationToken)
        {
            var service = await _serviceRepository.GetServiceByEncodedName(request.EncodedName)
                ?? throw new KeyNotFoundException($"Service with EncodedName {request.EncodedName} not found."); ;

            await _serviceRepository.DeleteAsync(service);
            return Unit.Value;
        }
    }
}
